added images for docs
