+++ src/lorepy/__init__.py
@@ -1,1 +1,2 @@
-from .lorepy import loreplot, uncertainty_plot
+from .lorepy import loreplot
+from .uncertainty import uncertainty_plot
