+++ src/lorepy/lorepy.py
@@ -1,7 +1,5 @@
 import pandas as pd
 from sklearn.linear_model import LogisticRegression
-from sklearn.model_selection import train_test_split
-from sklearn.utils import resample
 from pandas import DataFrame
 import numpy as np
 import matplotlib.pyplot as plt
@@ -107,106 +105,3 @@ def loreplot(
     ax.set_ylim(0, 1)
 
 
-def _get_uncertainty_data(
-    x: str,
-    X_reg,
-    y_reg,
-    x_range,
-    mode="resample",
-    jackknife_fraction: float = 0.8,
-    iterations: int = 100,
-    confounders=[],
-    clf=None,
-):
-    areas = []
-    for i in range(iterations):
-        if mode == "jackknife":
-            X_keep, _, y_keep, _ = train_test_split(
-                X_reg, y_reg, train_size=jackknife_fraction
-            )
-        elif mode == "resample":
-            X_keep, y_keep = resample(X_reg, y_reg, replace=True)
-        else:
-            raise NotImplementedError(
-                f"Mode {mode} is unsupported, only jackknife and resample are valid modes"
-            )
-
-        lg = LogisticRegression(multi_class="multinomial") if clf is None else clf
-        lg.fit(X_keep, y_keep)
-        new_area = _get_area_df(lg, x, x_range, confounders=confounders).reset_index()
-
-        areas.append(new_area)
-
-    long_df = pd.concat(areas).melt(id_vars=[x]).sort_values(x)
-
-    output = (
-        long_df.groupby([x, "variable"])
-        .agg(
-            min=pd.NamedAgg(column="value", aggfunc="min"),
-            mean=pd.NamedAgg(column="value", aggfunc="mean"),
-            max=pd.NamedAgg(column="value", aggfunc="max"),
-            low_95=pd.NamedAgg(column="value", aggfunc=lambda v: np.percentile(v, 2.5)),
-            high_95=pd.NamedAgg(
-                column="value", aggfunc=lambda v: np.percentile(v, 97.5)
-            ),
-            low_50=pd.NamedAgg(column="value", aggfunc=lambda v: np.percentile(v, 25)),
-            high_50=pd.NamedAgg(column="value", aggfunc=lambda v: np.percentile(v, 75)),
-        )
-        .reset_index()
-    )
-
-    return output
-
-
-def uncertainty_plot(
-    data: DataFrame,
-    x: str,
-    y: str,
-    x_range=None,
-    mode="resample",
-    jackknife_fraction=0.8,
-    iterations=100,
-    confounders=[],
-    colormap=None,
-    clf=None,
-):
-    X_reg, y_reg, r = _prepare_data(data, x, y, confounders)
-
-    if x_range is None:
-        x_range = r
-
-    plot_df = _get_uncertainty_data(
-        x,
-        X_reg,
-        y_reg,
-        x_range,
-        mode=mode,
-        jackknife_fraction=jackknife_fraction,
-        iterations=iterations,
-        confounders=confounders,
-        clf=clf,
-    )
-
-    categories = plot_df.variable.unique()
-
-    fig, axs = plt.subplots(ncols=len(categories), sharex=True, sharey=True)
-
-    cmap = plt.get_cmap("tab10") if colormap is None else colormap
-
-    for idx, category in enumerate(categories):
-        cat_df = plot_df[plot_df.variable == category]
-
-        axs[idx].fill_between(
-            cat_df[x], cat_df["low_95"], cat_df["high_95"], alpha=0.1, color=cmap(idx)
-        )
-        axs[idx].fill_between(
-            cat_df[x], cat_df["low_50"], cat_df["high_50"], alpha=0.2, color=cmap(idx)
-        )
-        axs[idx].plot(cat_df[x], cat_df["mean"], color=cmap(idx))
-        axs[idx].set_title(categories[idx])
-        axs[idx].set_xlabel(x)
-
-        axs[idx].set_xlim(*x_range)
-        axs[idx].set_ylim(0, 1)
-
-    return fig, axs
